segregate Docker-dependent E2E tests from standard suite
