+++ playwright-e2e.config.ts
@@ -35,6 +35,9 @@ export default defineConfig({
 				"**/gate-resign-cancel*",
 				"**/gates-api-heavy*",
 				"**/project-bugs*",
+				// Docker-dependent tests — run via test:manual instead
+				"**/sandbox-recovery-docker*",
+				"**/bg-process-sandbox-guard*",
 			],
 			workers: 4,
 		},
