+++ playwright-manual.config.ts
@@ -21,5 +21,13 @@ export default defineConfig({
 			name: "manual-integration",
 			testDir: "./tests/manual-integration",
 		},
+		{
+			name: "docker-e2e",
+			testDir: "./tests/e2e",
+			testMatch: [
+				"**/sandbox-recovery-docker*.spec.ts",
+				"**/bg-process-sandbox-guard*.spec.ts",
+			],
+		},
 	],
 });
