+++ tests/e2e/sandbox-recovery.spec.ts
@@ -1,17 +1,13 @@
 /**
- * E2E tests for sandbox container resilience — auto-recovery on container death.
+ * E2E tests for sandbox container resilience — process_exit event handling.
  *
  * Tests:
  * 1. `process_exit` event transitions session to `terminated` (no Docker needed)
- * 2. Container health monitor detects death and recreates container (Docker required)
- * 3. Sandbox sessions recover after container death (Docker required)
- * 4. Worktree branches intact after recovery (Docker required)
  *
- * The non-Docker test (Test 1) uses the in-process harness to directly access
- * the session manager and simulate RPC events. Docker-dependent tests skip when
- * Docker is unavailable.
+ * Docker-dependent tests (container health monitor, session recovery, worktree
+ * branches) have been moved to sandbox-recovery-docker.spec.ts and run via
+ * `npm run test:manual` instead.
  */
-import { execFileSync } from "node:child_process";
 import { test, expect } from "./in-process-harness.js";
 import {
 	apiFetch,
@@ -19,17 +15,7 @@ import {
 	connectWs,
 	waitForSessionStatus,
 	statusPredicate,
-	agentEndPredicate,
 } from "./e2e-setup.js";
-
-function isDockerAvailable(): boolean {
-	try {
-		execFileSync("docker", ["info"], { stdio: "ignore", timeout: 5000 });
-		return true;
-	} catch {
-		return false;
-	}
-}
 
 // ---------------------------------------------------------------------------
 // Test 1: process_exit event handling (no Docker required)
@@ -169,385 +155,5 @@ test.describe("process_exit event handling", () => {
 			conn.close();
 			await apiFetch(`/api/sessions/${id}`, { method: "DELETE" }).catch(() => {});
 		}
-	});
-});
-
-// ---------------------------------------------------------------------------
-// Docker-dependent tests: container health monitor and session recovery
-// ---------------------------------------------------------------------------
-
-test.describe("sandbox container recovery", () => {
-	const hasDocker = isDockerAvailable();
-
-	test.describe.configure({ mode: "serial" });
-
-	test("health monitor detects container death and recreates", async ({ gateway }) => {
-		test.skip(!hasDocker, "Docker not available");
-		test.setTimeout(120_000);
-
-		// 1. Get sandbox manager from gateway
-		const sm = gateway.sessionManager;
-		const sandboxManager = (sm as any).sandboxManager;
-
-		// If sandbox manager not available, skip
-		if (!sandboxManager) {
-			test.skip();
-			return;
-		}
-
-		// 2. Configure sandbox mode
-		const configResp = await apiFetch("/api/project-config", {
-			method: "PUT",
-			body: JSON.stringify({ sandbox: "docker" }),
-		});
-		expect(configResp.status).toBe(200);
-
-		// 3. Initialize sandbox for the default project
-		const projectId = (sm as any).pcm?.getDefaultProjectId?.() || "default";
-		try {
-			await sandboxManager.initForProject(projectId);
-		} catch (err) {
-			// If init fails (e.g., no git remote configured for sandbox), skip
-			test.skip();
-			return;
-		}
-
-		// 4. Get the current container ID
-		const sandbox = sandboxManager.get(projectId);
-		if (!sandbox) {
-			test.skip();
-			return;
-		}
-
-		let containerId: string;
-		try {
-			containerId = await sandbox.getContainerId();
-		} catch {
-			test.skip();
-			return;
-		}
-		expect(containerId).toBeTruthy();
-
-		// 5. Subscribe to health events and wait for recovery
-		const events: any[] = [];
-		const recoveryPromise = new Promise<void>((resolve, reject) => {
-			const timeout = setTimeout(() => reject(new Error("Recovery timeout (90s)")), 90_000);
-			sandbox.onHealthEvent((event: any) => {
-				events.push(event);
-				if (event.type === "container-recovered") {
-					clearTimeout(timeout);
-					resolve();
-				}
-			});
-		});
-
-		// 6. Kill the container
-		execFileSync("docker", ["rm", "-f", containerId], { timeout: 10_000, stdio: "ignore" });
-
-		// 7. Wait for recovery
-		await recoveryPromise;
-
-		// 8. Verify events sequence: container-died then container-recovered
-		expect(events.some(e => e.type === "container-died")).toBe(true);
-		expect(events.some(e => e.type === "container-recovered")).toBe(true);
-
-		// 9. Verify new container is running
-		const newContainerId = await sandbox.getContainerId();
-		expect(newContainerId).toBeTruthy();
-		expect(newContainerId).not.toBe(containerId);
-
-		const inspectResult = execFileSync("docker", [
-			"inspect", "--format", "{{.State.Running}}", newContainerId,
-		], { timeout: 5_000, encoding: "utf-8" }).trim();
-		expect(inspectResult).toBe("true");
-
-		// Cleanup
-		try {
-			await sandboxManager.shutdownAll();
-		} catch { /* best-effort */ }
-		await apiFetch("/api/project-config", {
-			method: "PUT",
-			body: JSON.stringify({ sandbox: "none" }),
-		}).catch(() => {});
-	});
-
-	test("sandbox sessions recover to idle after container death", async ({ gateway }) => {
-		test.skip(!hasDocker, "Docker not available");
-		test.setTimeout(180_000);
-
-		const sm = gateway.sessionManager;
-		const sandboxManager = (sm as any).sandboxManager;
-		if (!sandboxManager) {
-			test.skip();
-			return;
-		}
-
-		// 1. Configure sandbox
-		await apiFetch("/api/project-config", {
-			method: "PUT",
-			body: JSON.stringify({ sandbox: "docker" }),
-		});
-
-		const projectId = (sm as any).pcm?.getDefaultProjectId?.() || "default";
-		try {
-			await sandboxManager.initForProject(projectId);
-		} catch {
-			test.skip();
-			return;
-		}
-
-		const sandbox = sandboxManager.get(projectId);
-		if (!sandbox) { test.skip(); return; }
-
-		let containerId: string;
-		try {
-			containerId = await sandbox.getContainerId();
-		} catch {
-			test.skip();
-			return;
-		}
-
-		// 2. Create a sandboxed session
-		const createResp = await apiFetch("/api/sessions", {
-			method: "POST",
-			body: JSON.stringify({ cwd: nonGitCwd(), sandbox: true }),
-		});
-		expect(createResp.status).toBe(201);
-		const { id: sessionId } = await createResp.json();
-
-		await waitForSessionStatus(sessionId, "idle", 30_000);
-
-		// 3. Send a message to create chat history
-		const conn = await connectWs(sessionId);
-		try {
-			conn.send({ type: "prompt", text: "Hello from sandbox recovery test" });
-			await conn.waitFor(agentEndPredicate(), 30_000);
-		} finally {
-			conn.close();
-		}
-
-		await waitForSessionStatus(sessionId, "idle", 10_000);
-
-		// 4. Set up recovery watcher
-		const recoveryPromise = new Promise<void>((resolve, reject) => {
-			const timeout = setTimeout(() => reject(new Error("Session recovery timeout (90s)")), 90_000);
-			sandbox.onHealthEvent((event: any) => {
-				if (event.type === "container-recovered") {
-					clearTimeout(timeout);
-					resolve();
-				}
-			});
-		});
-
-		// 5. Kill the container
-		execFileSync("docker", ["rm", "-f", containerId], { timeout: 10_000, stdio: "ignore" });
-
-		// 6. Session should go terminated first (from process_exit)
-		await expect.poll(async () => {
-			const resp = await apiFetch(`/api/sessions/${sessionId}`);
-			const data = await resp.json();
-			return data.status;
-		}, { timeout: 30_000, intervals: [500] }).toBe("terminated");
-
-		// 7. Wait for container recreation + session recovery
-		await recoveryPromise;
-
-		// 8. Session should recover to idle
-		await expect.poll(async () => {
-			const resp = await apiFetch(`/api/sessions/${sessionId}`);
-			const data = await resp.json();
-			return data.status;
-		}, { timeout: 60_000, intervals: [1_000] }).toBe("idle");
-
-		// 9. Verify chat history is preserved — fetch messages
-		const messagesResp = await apiFetch(`/api/sessions/${sessionId}/messages`);
-		expect(messagesResp.status).toBe(200);
-		const messages = await messagesResp.json();
-		expect(messages.length).toBeGreaterThan(0);
-
-		// Cleanup
-		await apiFetch(`/api/sessions/${sessionId}`, { method: "DELETE" }).catch(() => {});
-		try { await sandboxManager.shutdownAll(); } catch { /* best-effort */ }
-		await apiFetch("/api/project-config", {
-			method: "PUT",
-			body: JSON.stringify({ sandbox: "none" }),
-		}).catch(() => {});
-	});
-
-	test("worktree branches intact after container recovery", async ({ gateway }) => {
-		test.skip(!hasDocker, "Docker not available");
-		test.setTimeout(180_000);
-
-		const sm = gateway.sessionManager;
-		const sandboxManager = (sm as any).sandboxManager;
-		if (!sandboxManager) { test.skip(); return; }
-
-		// 1. Configure sandbox
-		await apiFetch("/api/project-config", {
-			method: "PUT",
-			body: JSON.stringify({ sandbox: "docker" }),
-		});
-
-		const projectId = (sm as any).pcm?.getDefaultProjectId?.() || "default";
-		try {
-			await sandboxManager.initForProject(projectId);
-		} catch { test.skip(); return; }
-
-		const sandbox = sandboxManager.get(projectId);
-		if (!sandbox) { test.skip(); return; }
-
-		let containerId: string;
-		try {
-			containerId = await sandbox.getContainerId();
-		} catch { test.skip(); return; }
-
-		// 2. Create a sandboxed session (which gets a worktree)
-		const createResp = await apiFetch("/api/sessions", {
-			method: "POST",
-			body: JSON.stringify({ cwd: nonGitCwd(), sandbox: true }),
-		});
-		expect(createResp.status).toBe(201);
-		const { id: sessionId } = await createResp.json();
-
-		await waitForSessionStatus(sessionId, "idle", 30_000);
-
-		// 3. Get the session's branch and cwd
-		const session = sm.getSession(sessionId);
-		expect(session).toBeTruthy();
-		const sessionBranch = (session as any).branch ||
-			sm.getPersistedSession(sessionId)?.branch;
-		const sessionCwd = session!.cwd;
-
-		// 4. Set up recovery watcher
-		const recoveryPromise = new Promise<void>((resolve, reject) => {
-			const timeout = setTimeout(() => reject(new Error("Recovery timeout (90s)")), 90_000);
-			sandbox.onHealthEvent((event: any) => {
-				if (event.type === "container-recovered") {
-					clearTimeout(timeout);
-					resolve();
-				}
-			});
-		});
-
-		// 5. Kill the container
-		execFileSync("docker", ["rm", "-f", containerId], { timeout: 10_000, stdio: "ignore" });
-
-		// 6. Wait for recovery
-		await recoveryPromise;
-
-		// 7. Wait for session to recover
-		await expect.poll(async () => {
-			const resp = await apiFetch(`/api/sessions/${sessionId}`);
-			const data = await resp.json();
-			return data.status;
-		}, { timeout: 60_000, intervals: [1_000] }).toBe("idle");
-
-		// 8. Verify the worktree and branch exist in the new container
-		const newContainerId = await sandbox.getContainerId();
-		expect(newContainerId).toBeTruthy();
-
-		if (sessionCwd?.startsWith("/workspace-wt/")) {
-			// Verify the worktree directory exists
-			execFileSync("docker", [
-				"exec", newContainerId, "test", "-d", sessionCwd,
-			], { timeout: 5_000, encoding: "utf-8" });
-
-			// Verify the correct branch is checked out
-			if (sessionBranch) {
-				const branchOutput = execFileSync("docker", [
-					"exec", "-w", sessionCwd, newContainerId,
-					"git", "rev-parse", "--abbrev-ref", "HEAD",
-				], { timeout: 5_000, encoding: "utf-8" }).trim();
-				expect(branchOutput).toBe(sessionBranch);
-			}
-		}
-
-		// Cleanup
-		await apiFetch(`/api/sessions/${sessionId}`, { method: "DELETE" }).catch(() => {});
-		try { await sandboxManager.shutdownAll(); } catch { /* best-effort */ }
-		await apiFetch("/api/project-config", {
-			method: "PUT",
-			body: JSON.stringify({ sandbox: "none" }),
-		}).catch(() => {});
-	});
-
-	test("health monitor handles repeated container kills", async ({ gateway }) => {
-		test.skip(!hasDocker, "Docker not available");
-		test.setTimeout(240_000);
-
-		const sm = gateway.sessionManager;
-		const sandboxManager = (sm as any).sandboxManager;
-		if (!sandboxManager) { test.skip(); return; }
-
-		// 1. Configure sandbox
-		await apiFetch("/api/project-config", {
-			method: "PUT",
-			body: JSON.stringify({ sandbox: "docker" }),
-		});
-
-		const projectId = (sm as any).pcm?.getDefaultProjectId?.() || "default";
-		try {
-			await sandboxManager.initForProject(projectId);
-		} catch { test.skip(); return; }
-
-		const sandbox = sandboxManager.get(projectId);
-		if (!sandbox) { test.skip(); return; }
-
-		// --- First kill/recovery cycle ---
-		let containerId: string;
-		try {
-			containerId = await sandbox.getContainerId();
-		} catch { test.skip(); return; }
-
-		const firstRecovery = new Promise<void>((resolve, reject) => {
-			const timeout = setTimeout(() => reject(new Error("First recovery timeout")), 90_000);
-			const unsub = sandbox.onHealthEvent((event: any) => {
-				if (event.type === "container-recovered") {
-					clearTimeout(timeout);
-					unsub();
-					resolve();
-				}
-			});
-		});
-
-		execFileSync("docker", ["rm", "-f", containerId], { timeout: 10_000, stdio: "ignore" });
-		await firstRecovery;
-
-		// Verify first recovery succeeded
-		const afterFirst = await sandbox.getContainerId();
-		expect(afterFirst).toBeTruthy();
-		expect(afterFirst).not.toBe(containerId);
-
-		// --- Second kill/recovery cycle ---
-		const secondRecovery = new Promise<void>((resolve, reject) => {
-			const timeout = setTimeout(() => reject(new Error("Second recovery timeout")), 90_000);
-			const unsub = sandbox.onHealthEvent((event: any) => {
-				if (event.type === "container-recovered") {
-					clearTimeout(timeout);
-					unsub();
-					resolve();
-				}
-			});
-		});
-
-		execFileSync("docker", ["rm", "-f", afterFirst], { timeout: 10_000, stdio: "ignore" });
-		await secondRecovery;
-
-		// Verify second recovery also succeeded
-		const afterSecond = await sandbox.getContainerId();
-		expect(afterSecond).toBeTruthy();
-		expect(afterSecond).not.toBe(afterFirst);
-
-		const inspectResult = execFileSync("docker", [
-			"inspect", "--format", "{{.State.Running}}", afterSecond,
-		], { timeout: 5_000, encoding: "utf-8" }).trim();
-		expect(inspectResult).toBe("true");
-
-		// Cleanup
-		try { await sandboxManager.shutdownAll(); } catch { /* best-effort */ }
-		await apiFetch("/api/project-config", {
-			method: "PUT",
-			body: JSON.stringify({ sandbox: "none" }),
-		}).catch(() => {});
 	});
 });
